requirements.txt
flasgger==0.9.7.1
flask-marshmallow==0.15.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
mysqlclient==2.2.0
PyJWT==2.8.0
psycopg2==2.9.7
python-dotenv==0.21.0

